--- components/hooks/useModal.ts
@@ -1,22 +1,0 @@
-import { useRecoilState } from 'recoil';
-
-import { modalState } from '@recoil/atoms/modals';
-
-export default function useModal() {
-  const [modal, setModal] = useRecoilState(modalState);
-
-  const showModal = ({ modalType, modalProps }: any) => {
-    setModal({ modalType, modalProps });
-  };
-
-  const hideModal = () => {
-    setModal(null);
-  };
-
-  return {
-    modal,
-    setModal,
-    showModal,
-    hideModal,
-  };
-}
