+++ react-query/hooks/sentence/useRemoveSentence.ts
@@ -7,7 +7,7 @@ import axios from 'axios';
 import { useRecoilState } from 'recoil';
 
 import { queryKeys } from '@react-query/constants';
-import { useCustomToast } from '@components/hooks/useCustomToast';
+import { useCustomToast } from '@hooks/useCustomToast';
 import { contextState } from '@recoil/atoms/common';
 
 type Ids = {
